+++ src/main/java/Main.java
@@ -1,7 +1,6 @@
 
 public class Main {
     public static void main(String[] args) {
-        Dish dish = new Dish();
         Calculator calculator = new Calculator();
         calculator.peoplesCalculator();
         calculator.calculator2();
